requirements.txt
# Core dependencies
pathway[xpack-llm]
sentence-transformers
fastapi
uvicorn
python-dotenv
pydantic
requests
streamlit
openai

# Additional dependencies (these will be installed as needed by the core packages)
numpy
pandas
scikit-learn
torch
transformers
tqdm
